Just a short utility script for work (tableize).
On the dev servers at work, we have a script that's basically the same
as this one, but pretty well meant to be run only on those servers. It's
useful as an editor binding to be able to call. I just happen to want to
be able to call it from anywhere I'm set up, so I can edit from my own
workstation.