Fix bug which moved cursor when a pattern was searched for
Executing the match command non-silently would cause the cursor to move
to the beginning of the line. Unfortunately, we'd run it for every
change on the buffer, meaning that every time you typed, the cursor
would get moved. Bad for business.

Saving and restoring the view fixes it. Also threw in a keepjumps just
for safety.

It doesn't seem to matter if I use `silent` (with `redir`) here or not.
It's not working the way I want because, while I can pass a string in
a variable to `execute`, I cannot do so with `silent`. So I know of no
way to work around this problem other than saving and restoring the
view.