Use copy constructor for cache sentinel values
Construct and use the cache a little differently to fix a bug where the
cache always appeared to be not-stale. This was due to the fact that the
sentinel values were being updated along with the cached values because
Dictionaries are assigned by referenced (see `:help copy`).

Also, avoid updating the cache during evaluation until the last moment.
Not really any point there, aside from symmetry and lazier evaluation.
It does help performance in the cases of errors that might occur. It
might also prevent some glitches when evaluation gets interrupted. Yet
to be seen.