Fix toggling and display for searching large files
When initially toggling on match counting for a large file, the cache
still is populated with the cache's default values. Those are not
meaningful to display, so we should display a small indication that the
search is in progress.

Toggling should flip _on_ match counting first when a large file is
detected.