Attempt to fix a small corner case while searching large files
It's not entirely possible to fix this problem, but I did find a bug in
my script where I was checking for the length of a string being *less*
than zero rather than zero. That was, of course, incorrect.

I also commented out a line for later use for debugging, which I may
adapt later for use.

Finally, I added a small blurb acknowledging the issue to the script's
header.