Autocommit in psql for now (sigh)
For some stupid-ass reason, Rails' db:structure:load task calls
a PostgreSQL adapter method which is too stupid to know not to load the
local ~/.psqlrc file. Mine prevents autocommitting, as is wise, but the
adapter class blithely assumes this is not the case.

This has been fixed in a commit which has not been incorporated into
a stable release of Rails yet. So I'll have to disable it until this
happens.