Rework prompt appearance and use vcs_info
This is a very radical change in my Zsh prompt. It looks much different
(very powerline-appearing) and removes the dependencies on both the
oh-my-zsh git plugin that I'd copied over and the zsh colors plugin.

It's been entirely rewritten from scratch to use `vcs_info`. It's still
a work in progress, but it's gotten late, and it works, so I'm going to
stop for the night.