Make ImageMagick optional, shuffle several options
Use ImageMagick `convert` only if it exists. You can also disable it at
runtime, even if you have it installed. If you try to enable it without
having it, you get a warning. You can disable the warning, though, for
some reason.

As a side-effect of all this, the output filename is now unadorned. This
also means that the input and output filenames must be listed in that
order. Otherwise, ordering is still unimportant.